Initial simple move to fetch Constituencies via an External CF
All this currently does is read from the internal CustomFieldValues
table.  It provides the scaffolding that Constituency needs to limit
available values on a per user / ticket basis.

EditCustomFieldConstituency just calls through to a simple
CustomFieldSelect for now.

Removes EditConstituency which was previously used to generate the
trimmed list of Constituencies for certain users.

Mild shenanigans are required in initialdata to ensure that the
CustomFieldValue is set prior to make initdb (or
RT::Test::bootstrap_db).  Neither of these codepaths loads plugins
before running initialdata.  Doing something like Assets and require'ing
RT::IR will cause problems with all the various caches set up in the
package.  Instead, we just continue the trend of dummy lifecycles with
manual frobbing of CustomFieldValuesSources and ensuring that
RT::CustomFieldValues::Constituency is loaded so that our ValuesClass is
obeyed (and RenderTypes are found/validated).