Use \b as a stronger limiter for before IP addresses
Previously, "doesn't come after any possible IP characters" was the
limit.  For IPv6 addresses, however, this caused unwanted behavior:

   dead:beef::       => dead:beef::
   :dead:beef::      => ead:beef::
   qdead:beef::      => dead:beef::
   meathead:beef::   => ead:beef::
   IPv6:dead:beef::  => 6:dead:beef::
   wrong:dead:beef:: => ead:beef::

There is no harm in allowing an IPv6 address to follow after a colon.
Additionally, to ensure that "IPv6:dead:beef::" does not merely take the
"6", limit both IPv4 and IPv6 addresses to coming at a word boundary.

With these changes, we obtain:

    dead:beef::       => dead:beef::
    :dead:beef::      => dead:beef::
    qdead:beef::      => beef::
    meathead:beef::   => beef::
    IPv6:dead:beef::  => dead:beef::
    wrong:dead:beef:: => dead:beef::

All of which are more expected.  Note that this also causes
"base64.1.2.3" to no longer match as an IPv4 address; this is considered
to be an appropriate change.

Fixes: I#30643