Remove stale callback on upgrade from RTIR 3.2.0
RTIR 4.0 doesn't need the callback
html/Callbacks/RTIR/Elements/EditCustomField/EditComponentName any
longer, and so it was deleted from the package. But upgrades from 3.2.0
would of course keep that file around, which caused errors in the admin
UI for ticket custom fields:

[10257] [Fri Jun 17 22:01:02 2016] [error]: Can't call method "QueueObj" on an undefined value at /opt/rt4/local/plugins/RT-IR/html/Callbacks/RTIR/Elements/EditCustomField/EditComponentName line 51.

This uses a new (generic) feature added to Module::Install::RTx for the
purpose of removing this file on upgrade.

Fixes: I#32100