Remove Type arg from call to UpdateData template
The arg Type => Incident was being passed to the UpdateData
template, but that template didn't have Type in the args list
and wasn't using the value. However, it ended up being passed
along to RT's MessageBox template in %ARGS where it overrode
the Type arg there, which is used to indicate the Content Type for
the update. This resulted in the plain text comment box being displayed
even when rich text was configured in RT.

Fixes: I#32166