Fix Show email link under "immediate" show history
Without this fix, we were generating links to
/RTIR/Ticket/ShowEmailRecord.html which does not exist. Now we correctly
link to /Ticket/ShowEmailRecord.html.

This also fixes the links to attachments in the transaction log (which
were linking to the nonexistent /RTIR/Ticket/Attachment/…)

This new PathPrefix value corresponds to the one we provide in the
ExtraShowHistoryArguments callback, which is why it works for the
various deferred show history preferences.

Fixes: T#183105