Redirect to the correct display page from Update.html
Update.html executes Display.html to process ticket updates
similar to the flow in RT. However, RTIR has a custom Display.html
page for incidents. The default RTIR/Display.html checks for custom
display pages and issues a redirect to RTIR/Incident/Display.html
for incident updates. This results in a 302 back to the browser which
generates a subsequent GET request to the correct Display.html page.
This moves all of the previous POST args to the URI string, which
can result in Request-URL string too long errors.

Move the custom page detection logic to Update.html so it can
execute the correct page without the extra redirect.