Discover Lifecycle if not passed to load proper Format
RTIR supports many custom search result formats, including
a format for each special lifecycle type. The results page
was checking for a custom type format based on the Lifecycle, but
in some cases the lifecycle isn't passed as part of the Query.
Discover it based on the query so custom search result formats
can be loaded and used.