Remove extra table row at bottom of RTIR ticket display
RTIR ticket display had an extra table row at the bottom of
the page such that if a ticket had a large number of custom
fields, the More about requestors box would be pushed way down
the page, creating a large amount of empty white space above
and to the left of it.

Move the components from the extra row into the same positions in
the row above, allowing all components to display below the ones
above them.