Avoid test failure with 4.4.2's new Reports menu
Previously, when this test clicked the "Reports" menu it expected to
land on a list of Incident Reports. Now, when it clicks the menu labeled
"Reports" it ends up with a list of possible charts.

follow_link_ok supports multiple parameters to better narrow down which
link you intended to follow, so also specify the url must match /RTIR/

Fixes: T#187988