Scope side by side layout to still allow the single column option
RT added the side by side layout in the 3.8 series as the default,
but still allowed the single column with a configuration option and
a user preference. In RT, 4c22c70dc moved this to a css class to
accomodate squished css files. Mirror this change in RTIR to still
allow the single column option for Incident and Investigation
create pages.