Drive select for state to status change from tickets
The 2.9.0 upgrade script searches for the older state custom
field and uses it to set status appropriately. The previous
search worked from existing states in the state custom field, but
state information is retained by custom fields when a ticket is
moved to a different queue. The query would therefore find a state
for an IR (new), for example, even if the ticket had been moved to
the Incidents queue and there had a state of open. After an upgrade,
the Incident ticket would have a status of new, which is incorrect and
also not a valid status in the incidents lifecycle.

Re-arrange the query to work from tickets rather than states to only
pull state data and update status using current ticket information.