Lifecycle is needed to build ResultPage and BaseQuery from ARGS
Here is a failure case without it: to link IRs to an incident(via "Link"
on incident display page), the "Edit Search" link has args "id" and
"Lifecycle", which is set to the incident id and "incident_reports",
respectively.

After any action in the builder(e.g. "Add these terms and Search"),
"Lifecycle" arg is lost, so ResultPage and BaseQuery relying on it are
wrong, which results in wrong links in page menu and also wrong page
title.