Stricter IPv6 regex by adding more limitations to the surroundings.
We want a clear break between IPv6 and its surroundings to not extract
"::Add" from "Scan::Address_Scan". Because as a valid IPv6 itself,
"::Add" is definitely not meant to be one in that string.

So the general rule is: surroundings shouldn't contain common characters
valid to IPv6(i.e. "[0-9a-fA-F:.]"), and there should be a word break,
thus the not allowed characters become "[0-9a-zA-Z:.]".

An exception is trailing dot, which is commonly used as the end of a
sentence. Like how we handle it in IPv4 regex, it's allowed as long as
it's not directly followed by suspicious "[0-9a-zA-Z:]".