Massage grouped CF inputs too on "Create Incident and run Investigation"
On "Create Incident and run Investigation", the form contains CF values
for both queues. We need to filter out CF inputs that are not applied to
the current queue, so on CreateTicket call, RT could get a clean %ARGS
and not issue the error log of:

    Invalid context object RT::Queue (5) for CF 2; skipping CF

Previously it doesn't consider CF groupings, this commit fixes this.