Pass Requestors through queue selection modal if provided
ebe8874d introduced the new queue selection modal which requires some
values intended for the create page to be written as hidden fields so
they are passed on when submitted.

This change adds Requestors to support the Investigate to
link. Previously the "Investigate to" email addresses were being lost
on the queue modal and not appearing in the Correspondents field.