Handle emails with +tags in 'Investigate to' MakeClicky links
Email addresses with +tags were not URI-escaped in the RTIR MakeClicky
callback for 'Investigate to' links. Then, the CreateInRTIRQueueModal
would URI-unescape the Requestors GET parameter, replacing + with a
space in the hidden input field. The modal passed that incorrect value
to the RTIR investigation creation page.

This change URI-escapes the email address in the RTIR MakeClicky
callback before inserting in to the 'InvestigateTo' link. That allows
the modal to unescape it to a correct value, and pass the correct
value to the investigation creation page.