Migrate modals to new elevator themes
Although RT no longer uses modals for the create pages, the
functionality in RTIR is slightly different.  Since there can be
more than one queue (with incidents for example), the ability to
select the queue on create is still required.  For now, the logic
within rtir.js still serves the ability to select which queue, but
will later be changed to add the select dropdown directly to the
create pages.