Check individual linked incidents on split page
The previous is_deeply call failed when running on the Debian
test Docker container because it appears to display the second
linked incident twice. This shows up in the dumped HTML and in the
params returned in the test.

Unable to reproduce this manually in a running RTIR, this update converts the
test to an affirmative test for the links. However the test may be
revealing a bug that only happens in certain circumstances, so leaving
a comment for further investigation.