Add RTIR_DefaultQueue option
Queue is now selected on the ticket create page, so add a
new DefaultQueue option for RTIR create pages. This allows the
RT create page to have a separate default queue if desired.

As part of this change, convert the direct META setting to
use the AddOption method and provide these options on the
user preferences page.