Remove default setting of DefaultQueue for RT
With RT's DefaultQueue set to an RTIR queue, users have no way of
getting to the regular RT ticket create page because clicking on the
Create new ticket button will allows redirect to the RTIR create page.

Remove this default setting and note it in the UPGRADING file. The new
RTIR_DefaultQueue option is now available to provide similar
functionality for RTIR top action of "Create new ticket".