Update test for modal removal
The functionality tested in modal.t is still valueable, although
not correctly named since removing the create in queue modal. This
commit renames the test as well as updates the tests to verify the
queue select options on the Create pages instead.

Additionally, the regexs within the tests were slightly loosened to
more easily allow for pre-selected options in the dropdown.