Allow investigation queue to be selected
This commit updates the investigation on create element to allow
changing the investigation queue.  Prior to this the queue was
tied to the selected incident and couldn't be changed.

Both selected incident and selected investigation now persist
through the reload by reading the InvestigationQueue arg on post.

If the user can't load the Queue then load one they can
SeeQueue and CreateTicket for.