Link to RT ticket create on RTIR create pages
It's possible that users couldn't access RT ticket create page, when
their default queue is an RTIR one, in which case RT will redirect to
RTIR create pages automatically.

This commit adds a link to RT ticket create page with a non-RTIR queue
specified, to get around this edge issue.