Don't redirect to RTIR create on RT create page if there is no default queue
When there is no default queue, RT defaults to the first queue current
user can create in alphabetic order, which could be an RTIR one. In such
case, it's more correct to stick to RT create page, where user could
choose from all the queues. So here we default to the first non-RTIR
queue instead.