Remove the confusing incident Reply page menu
The incident reply page, accessed via "Reply to Reporters" or
"Reply to All" had a "Reply" entry in the submenu. This menu
was originally to allow users to navigate back to the reply
page after using "Edit Search". Users can now do this by
clicking on "Show Results" after modifying their search.

Showing it on the reply page itself is confusing since it
takes you to the page you are already on.