Add ARG for /RTIR/Elements/ShowIncidents to add classes when in a form
When values are inside a form, the alignment of the label and value
are adjusted with CSS. This means when we display the results of
ShowIncidents during inline edit we need to add these CSS classes.

This also fixes the alignment issue on the Edit.html page when showing
no linked incidents (none).