Fix squelching functionality on update page
We need to keep track of submitted squelching info, so the state is not
lost when user submits the form to refresh, e.g. via "Select an Article
to include".

Note that RT's update page calls both "ShowSimplifiedRecipients" and
"PreviewScrips". To avoid duplication, "ShowSimplifiedRecipients"
renders the hidden input only if current user lacks ShowOutgoingEmail
right by default. As we don't use "PreviewScrips" in RTIR, we need to
force render the input in "ShowSimplifiedRecipients"