Fix Query of "Edit Search" menu on merge pages
Earlier $search_arguments->() already gets the correct Query from
"SearchArgs" param of "/Elements/Tabs", so there is no need to
re-generate, not mentioning that the generation is wrong.

This commit fixes the issue that Query param of "Edit Search" on Merge
pages was empty, which was introduced when we migrated to RT search
pages in 5.0.