Convert Dockerfile options to build arguments
In order to support different database configurations, it needs to be
possible to set these variables at image build time, not just container
runtime. Converting these options to build arguments is the most natural
way to do that. We capture the variables that the RTIR test runner
actually needs at the end of the build.