Add tooltip to select incident text input if it's below the label
The "Incident" label has the tooltip, but when the text input is not in the
same row, people might be confused, especially if the selectable incident
list is long. This commit adds another similar tooltip when the label and
text input are not aligned.

Note that the tooltip doesn't contain the "Required" or "Optional" info,
because it depends on if there is an incident selected or not. It's too
much to dynamicaly update the tooltip depending on other inputs, especially
that the tooltip is hidden by default, so people might not even notice
the change. Keeping this part in the main "Incident" label seems enough.