Replace ToolFormIframe by ToolFormExternal
ToolFormIframe was created with the concept of opening useful tools such
as Siteadvisor and CVE in an iframe, helping to debug and search ticket
informations. Since the adoption of X-Frame-Options by most websites,
this feature is no longer useful. This commit keeps the same behavior
but opens the tools in a new tab of the browser.