Fix custom field grouping by queue/category
Since RT 5.0.3 it is possible to create custom grouping categories per
queue. RTIR however was overriding part of the code that was responsible
for this feature to work.

This commit adjust the code to use the new feature and makes RTIR not
break the custom field grouping by queue/category.