Move Menu() calls to the new PrivilegedMainNav callback
RT 6 doesn't always refresh the top menu, so Menu is no longer
available on every call. Move the top menu items to the new
callback that will be called at least on the initial page load.

Search menus are intentionally removed because both Edit Search
and Show Results are available in the page menu when working
with a search, so they are not needed in the top level RTIR
menu.