Update formatting for lookup tools for RT 6
Also convert the traceroute system call to use IPC::Run3. The
previous implementation suggested it might stream results to the
browser one line at a time, but in testing it seemed it waited
until all results were returned. Since that handle approach
didn't have that advantage, switch to the utility function used
in RT for calls to system tools.