Update Reply All pages to improve layout and appearance
The Reply All page is called from individual incidents and from
Bulk Abandon. Both offered time inputs, but time has not worked for a
long time from Bulk Abandon because it loops through many tickets and
any time is lost in the process. It's also not clear where the time should
be recorded when there are many incidents. The time input is now
shown just for the individual incident update.

Additional updates work to update labels to RT 6 styling and try to
compress other inputs a bit.