Use different page layouts on Incidents based on Classification
RT 6 allows you to show different page layouts based on the value
of a custom field. This was a very common request for RTIR because
users want to show different fields on the Incident page based on
the type of incident, which is identified by the Classification
custom field.

Add a basic configuration that creates all of the elements to
enable these new page layouts.